+++ module/LearnZF2Ajax/test/Bootstrap.php
@@ -11,8 +11,8 @@ error_reporting(E_ALL | E_STRICT);
 chdir(__DIR__);
 
 /**
-* Test bootstrap, for setting up autoloading
-*/
+ * Test bootstrap, for setting up autoloading
+ */
 class Bootstrap
 {
     protected static $serviceManager;