+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ### Added
 
-- [67: Added Doctrine configuration support](https://github.com/sitrunlab/LearnZF2/pull/67)
+- [67: Doctrine configuration support](https://github.com/sitrunlab/LearnZF2/pull/67)
 - [69: Db usage for module list](https://github.com/sitrunlab/LearnZF2/pull/69)
+- [69: Pagination module](https://github.com/sitrunlab/LearnZF2/pull/70)
 
 
 ## 0.0.2 - 2014-12-03