+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - [87: set title prepend per module and add title in link for seo](https://github.com/sitrunlab/LearnZF2/pull/87)
 - [89: Fixes bug : using viewmodel->terminate() to check whenever view has layout](https://github.com/sitrunlab/LearnZF2/pull/89)
 - [90: hotfix : restrict other "in db" modules to be download-able](https://github.com/sitrunlab/LearnZF2/pull/90)
+- [91: remove "ant" on .travis.yml, use "parallel" instead](https://github.com/sitrunlab/LearnZF2/pull/91)
 
 ## 0.0.2 - 2014-12-03